Server commit

+++ Server/Data/Ticket.cs
@@ -3,7 +3,7 @@
 public sealed class Ticket
 {
     public int Id { get; init; }
-    public required sbyte NumberOfPassengers { get; init; }
+    public required byte NumberOfPassengers { get; init; }
     public required int TripId { get; init; }
     public Trip? Trip { get; init; }
     public required int UserId { get; init; }
@@ -12,5 +12,5 @@ public sealed class Ticket
     public City? StartCity { get; init; }
     public required int EndCityId { get; init; }
     public City? EndCity { get; init; }
-    public required int TotalPrice { get; init; }
+    public required decimal TotalPrice { get; init; }
 }
